feat(core): support relevant snippet from changed files
